pyproject.toml
[project]
name = "ae2-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-timescaledb>=0.4.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.0",
    "pytest-env>=1.1.5",
    "ruff>=0.11.13",
    "ty>=0.0.1a10",
]
flask = [
    "alembic>=1.16.2",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
]
socket = [
    "keepalive-socket>=0.0.1",
]

[tool.uv]
default-groups = "all"
